moved playback section to Home as a child to provider
