+++ spotify-ui/src/components/pages/Home.tsx
@@ -1,14 +1,13 @@
-import React, {useEffect} from 'react'
-import SpotifyPlayback from '../playback/SpotifyPlayback'
-import PlaybackContextProvider from '../playback/PlaybackContext'
+import React from 'react'
+import PlaybackSection from '../playback'
+import PlaybackContextProvider from '../playback/context/PlaybackContext'
 export default function Home() {
-  useEffect(() => {
-    console.log('inside seths butt')
-  }, [])
-  console.log('outside seths butt')
+  
   return (
     <div>
-      <PlaybackContextProvider/>
+      <PlaybackContextProvider>
+        <PlaybackSection />
+      </PlaybackContextProvider>
     </div>
   )
 }
